Clean & change object color in function of facts
